pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spy2ytb"
version = "0.1.0"
description = "Download Spotify playlists from YouTube with full metadata"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mutagen>=1.47.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "spotipy>=2.23.0",
    "yt-dlp>=2023.0.0",
]

[project.scripts]
spy2ytb = "spy2ytb.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
